+++ build.zig
@@ -252,9 +252,6 @@ pub fn build(b: *Build) !void {
             "mir",
         },
     ) catch unreachable;
-    if (build_options.mimalloc) {
-        sys_libs.append("mimalloc") catch unreachable;
-    }
 
     includes.appendSlice(&[_][]const u8{
         "/usr/local/include",
@@ -269,6 +266,7 @@ pub fn build(b: *Build) !void {
     }) catch unreachable;
 
     const lib_pcre2 = try buildPcre2(b, target, build_mode);
+    const lib_mimalloc = if (build_options.mimalloc) try buildMimalloc(b, target, build_mode) else null;
 
     // If macOS, add homebrew paths
     if (builtin.os.tag == .macos) {
@@ -351,6 +349,12 @@ pub fn build(b: *Build) !void {
     lib.addOptions("build_options", build_options.step(b));
 
     lib.linkLibrary(lib_pcre2);
+    if (lib_mimalloc) |mimalloc| {
+        lib.linkLibrary(mimalloc);
+        if (lib.target.getOsTag() == .windows) {
+            lib.linkSystemLibrary("bcrypt");
+        }
+    }
     // So that JIT compiled function can reference buzz_api
     exe.linkLibrary(lib);
 
@@ -431,6 +435,12 @@ pub fn build(b: *Build) !void {
         }
         std_lib.main_mod_path = .{ .path = "src" };
         std_lib.linkLibrary(lib_pcre2);
+        if (lib_mimalloc) |mimalloc| {
+            std_lib.linkLibrary(mimalloc);
+            if (std_lib.target.getOsTag() == .windows) {
+                std_lib.linkSystemLibrary("bcrypt");
+            }
+        }
         std_lib.linkLibrary(lib);
         std_lib.addOptions("build_options", build_options.step(b));
 
@@ -474,6 +484,13 @@ pub fn build(b: *Build) !void {
     }
     if (build_options.needLibC()) {
         tests.linkLibC();
+    }
+    tests.linkLibrary(lib_pcre2);
+    if (lib_mimalloc) |mimalloc| {
+        tests.linkLibrary(mimalloc);
+        if (tests.target.getOsTag() == .windows) {
+            tests.linkSystemLibrary("bcrypt");
+        }
     }
     tests.addOptions("build_options", build_options.step(b));
 
@@ -548,6 +565,78 @@ pub fn buildPcre2(b: *Build, target: std.zig.CrossTarget, optimize: std.builtin.
     lib.installHeader("vendors/pcre2/src/pcre2.h.generic", "pcre2.h");
     lib.linkLibC();
     b.installArtifact(lib);
+
+    return lib;
+}
+
+pub fn buildMimalloc(b: *Build, target: std.zig.CrossTarget, optimize: std.builtin.OptimizeMode) !*Build.Step.Compile {
+    const lib = b.addStaticLibrary(
+        .{
+            .name = "mimalloc",
+            .target = target,
+            .optimize = optimize,
+        },
+    );
+
+    lib.addIncludePath(.{ .path = "./vendors/mimalloc/include" });
+    lib.linkLibC();
+
+    if (lib.target.getOsTag() == .macos) {
+        var macOS_sdk_path = std.ArrayList(u8).init(b.allocator);
+        try macOS_sdk_path.writer().print(
+            "{s}/usr/include",
+            .{
+                (std.ChildProcess.exec(.{
+                    .allocator = b.allocator,
+                    .argv = &.{
+                        "xcrun",
+                        "--show-sdk-path",
+                    },
+                    .cwd = b.pathFromRoot("."),
+                    .expand_arg0 = .expand,
+                }) catch {
+                    std.debug.print("Warning: failed to get MacOSX sdk path", .{});
+                    unreachable;
+                }).stdout,
+            },
+        );
+
+        lib.addSystemIncludePath(.{ .path = macOS_sdk_path.items });
+        // Github macos-12 runner (https://github.com/actions/runner-images/blob/main/images/macos/macos-12-Readme.md).
+        lib.addSystemIncludePath(.{ .path = "/Applications/Xcode_14.0.1.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/include" });
+        lib.addSystemIncludePath(.{ .path = "/Library/Developer/CommandLineTools/SDKs/MacOSX14.0.sdk/usr/include" });
+        lib.addSystemIncludePath(.{ .path = "/Library/Developer/CommandLineTools/SDKs/MacOSX13.3.sdk/usr/include" });
+        lib.addSystemIncludePath(.{ .path = "/Library/Developer/CommandLineTools/SDKs/MacOSX12.3.sdk/usr/include" });
+        lib.addSystemIncludePath(.{ .path = "/Library/Developer/CommandLineTools/SDKs/MacOSX12.1.sdk/usr/include" });
+    }
+
+    lib.addCSourceFiles(
+        .{
+            .files = &.{
+                "./vendors/mimalloc/src/alloc-aligned.c",
+                "./vendors/mimalloc/src/alloc.c",
+                "./vendors/mimalloc/src/arena.c",
+                "./vendors/mimalloc/src/bitmap.c",
+                "./vendors/mimalloc/src/heap.c",
+                "./vendors/mimalloc/src/init.c",
+                "./vendors/mimalloc/src/options.c",
+                "./vendors/mimalloc/src/os.c",
+                "./vendors/mimalloc/src/page.c",
+                "./vendors/mimalloc/src/random.c",
+                "./vendors/mimalloc/src/segment-map.c",
+                "./vendors/mimalloc/src/segment.c",
+                "./vendors/mimalloc/src/stats.c",
+            },
+            .flags = if (lib.optimize != .Debug)
+                &.{
+                    "-DNDEBUG=1",
+                    "-DMI_SECURE=0",
+                    "-DMI_STAT=0",
+                }
+            else
+                &.{},
+        },
+    );
 
     return lib;
 }