+++ CHANGELOG.md
@@ -43,7 +43,8 @@
 - `list.append` does not return the appended value anymore
 - Generic types syntax changed from `myFunction(<K,V>, ...)` to `myFunction::<K,V>(...)`
 - Nullable object fields and nullable variables have a `null` initial value if none is provided
-- Migrated to pcre2 which is now a submodule and built by build.zig
+- Migrated to pcre2 which is now a submodule built by build.zig
+- Mimalloc is now a submodule built by build.zig
 
 ## Fixed